7/18 : 1주차 과제 정답 중 오류 수정

+++ week1/hwanswer/triangle_answer.c
@@ -2,10 +2,10 @@
 
 int main()
 {
-    int x1, x2, x3, y1, y2, y3;
-    scanf("%d %d", &x1, &y1);
-    scanf("%d %d", &x2, &y2);
-    scanf("%d %d", &x3, &y3);
+    double x1, x2, x3, y1, y2, y3;
+    scanf("%lf %lf", &x1, &y1);
+    scanf("%lf %lf", &x2, &y2);
+    scanf("%lf %lf", &x3, &y3);
 
     if(x1 == x2 && x2 == x3)
     {
